Fix resize window display
