+++ client/src/components/features/Button.jsx
@@ -1,16 +1,13 @@
-import { useState } from "@hookstate/core";
+// import { useState } from "@hookstate/core";
 import React from "react";
-import globalstore from "../../Store";
+// import globalstore from "../../Store";
 
 const Button = ({ children, onClick }) => {
-  const store = useState(globalstore);
-  console.log(store.exportbuttondistabled.get());
+  // const store = useState(globalstore);
+  // console.log(store.exportbuttondistabled.get());
+  // const distable = store.exportbuttondistabled.get();
   return (
-    <button
-      className="Btn"
-      onClick={onClick}
-      // distabled={!!store.exportbuttondistabled.get()}
-    >
+    <button className="Btn" onClick={onClick}>
       {children}
     </button>
   );
